refactor(multiselect): seperate styles and create a new context
