+++ packages/chakra-components/src/MultiSelect/index.js
@@ -1,12 +1,9 @@
-import React, { forwardRef } from "react";
-import { get } from "styled-system";
-import {
-  PseudoBox,
-  Icon,
-  useFormControl,
-  useTheme,
-  useColorMode,
-} from "@chakra-ui/core";
+import React, { forwardRef, createContext } from "react";
+import { PseudoBox, Icon, useFormControl } from "@chakra-ui/core";
+
+import { useMultiSelectStyle } from "./styles";
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 const Input = forwardRef((props, ref) => {
   const {
@@ -60,26 +57,28 @@ Input.defaultProps = {
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+const MultiSelectContext = createContext();
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 export const MultiSelect = ({
   size,
   focusBorderColor,
   errorBorderColor,
   ...rest
 }) => {
-  const theme = useTheme();
-  const { colorMode } = useColorMode();
-  const wrapperStyleProps = wrapperStyle({
-    theme,
-    colorMode,
+  const context = {};
+
+  const styleProps = useMultiSelectStyle({
     size,
     focusBorderColor,
     errorBorderColor,
   });
 
   return (
-    <>
+    <MultiSelectContext.Provider value={context}>
       <PseudoBox pos="relative">
-        <PseudoBox {...wrapperStyleProps} {...rest}>
+        <PseudoBox {...styleProps} {...rest}>
           <PseudoBox
             position="relative"
             display="flex"
@@ -143,9 +142,9 @@ export const MultiSelect = ({
             <Icon name="warning" />
           </PseudoBox>
         </PseudoBox>
-        <input type="hidden" name="color" value="red" />
+        <HiddenInput />
       </PseudoBox>
-    </>
+    </MultiSelectContext.Provider>
   );
 };
 
@@ -159,96 +158,8 @@ MultiSelect.defaultProps = {
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-const wrapperStyle = ({
-  theme,
-  colorMode,
-  size,
-  focusBorderColor,
-  errorBorderColor,
-}) => {
-  const bg = { light: "white", dark: "whiteAlpha.100" };
-  const borderColor = { light: "inherit", dark: "whiteAlpha.50" };
-
-  const baseProps = {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "space-between",
-    flexWrap: "wrap",
-    position: "relative",
-    cursor: "default",
-    transition: "all 0.2s",
-    outline: "none",
-    appearance: "none",
-    bg: bg[colorMode],
-    border: "1px",
-    borderColor: borderColor[colorMode],
-  };
-
-  const inputSizes = {
-    lg: {
-      fontSize: "lg",
-      height: 12,
-      rounded: "md",
-    },
-    md: {
-      fontSize: "md",
-      height: 10,
-      rounded: "md",
-    },
-    sm: {
-      fontSize: "sm",
-      height: 8,
-      rounded: "sm",
-    },
-  };
-
-  const sizeProps = inputSizes[size];
-
-  /**
-   * styled-system's get takes 3 args
-   * - object or array to read from
-   * - key to get
-   * - fallback value
-   */
-  const hoverColor = { light: "gray.300", dark: "whiteAlpha.200" };
-  const _focusBorderColor = get(
-    theme.colors,
-    focusBorderColor,
-    focusBorderColor, // If color doesn't exist in theme, use it's raw value
-  );
-  const _errorBorderColor = get(
-    theme.colors,
-    errorBorderColor,
-    errorBorderColor,
-  );
-
-  const interactionProps = {
-    _hover: {
-      borderColor: hoverColor[colorMode],
-    },
-    _disabled: {
-      opacity: "0.4",
-      cursor: "not-allowed",
-    },
-    _focus: {
-      zIndex: 1,
-      borderColor: _focusBorderColor,
-      boxShadow: `0 0 0 1px ${_focusBorderColor}`,
-    },
-    _invalid: {
-      borderColor: _errorBorderColor,
-      boxShadow: `0 0 0 1px ${_errorBorderColor}`,
-    },
-    _readOnly: {
-      bg: "transparent",
-      boxShadow: "none !important",
-      userSelect: "all",
-    },
-  };
-
-  return {
-    ...baseProps,
-    ...sizeProps,
-    ...interactionProps,
-  };
+const HiddenInput = () => {
+  return <input type="hidden" name="color" value="red" />;
 };
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
